refactor: search.h

+++ src/search.c
@@ -47,7 +47,7 @@ static void search_next_forward(ui_t *ui, Fm *fm, bool inclusive)
 
 	Dir *dir = fm_current_dir(fm);
 	highlight(ui, NULL);
-	for (int i = inclusive ? 0 : 1; i < dir->length; i++) {
+	for (uint16_t i = inclusive ? 0 : 1; i < dir->length; i++) {
 		if (strcasestr(dir->files[(dir->ind + i) % dir->length]->name, search_string)) {
 			fm_move_to_ind(fm, (dir->ind + i) % dir->length );
 			return;
@@ -63,7 +63,7 @@ static void search_next_backwards(ui_t *ui, Fm *fm, bool inclusive)
 
 	Dir *dir = fm_current_dir(fm);
 	highlight(ui, NULL);
-	for (int i = inclusive ? 0 : 1 ; i < dir->length; i++) {
+	for (uint16_t i = inclusive ? 0 : 1 ; i < dir->length; i++) {
 		if (strcasestr(dir->files[(dir->length + dir->ind - i) % dir->length]->name, search_string)) {
 			fm_move_to_ind(fm, (dir->length + dir->ind - i) % dir->length );
 			return;
